+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Bug #15: Update PHPStan annotations in `DataProvider` and `HasDataTest` classes for improved type clarity (@terabytesoftw)
 - Enh #16: Enhance support for `Stringable` interface types across `HasClass`, `HasData`, `HasStyle`, `HasTitle` and update related tests (@terabytesoftw)
 - Bug #17: Add cases for `<details>`, `<dialog>`, and `<menu>` HTML tags in `Block` enum (@terabytesoftw)
+- Enh #18: Add `HasAccesskey` trait with `accesskey()` method and tests (@terabytesoftw)
 
 ## 0.2.0 December 18, 2025